fixed casting in celli, cellj

+++ grid.cc
@@ -16,11 +16,11 @@ Grid::~Grid(){
 // Given r returns i coordinate of containing grid cell
 // returned as double so user can get position in cell
 double Grid::celli(double r) {
-    return (int) (r - R0)*(1.0/dr);
+    return (r - R0)*(1.0/dr);
 }
 
 // Given z returns j coordinate of containing grid cell
 // returned as double so user can get position in cell
 double Grid::cellj(double z) {
-    return (int) (z - z0)*(1.0/dz);
+    return (z - z0)*(1.0/dz);
 }
